+++ libraries/FreeIMU/examples/FreeIMU_raw/FreeIMU_raw.pde
@@ -13,7 +13,7 @@
 #include <Wire.h>
 #include <SPI.h>
 
-int raw_values[11];
+int16_t raw_values[11];
 char str[512];
 float val[9];
